legendplot: Standard and 'rgl' Plots with Legends

This package provides tools to combine standard R plots or 'rgl' 3D plots with a legend, facilitating the creation of composite figures that mix 2D or 3D visualizations with a categorical or continuous legend. For more information visit https://rubenfcasal.github.io/legendplot/articles/legendplot.html.

See also

Author

Maintainer: Ruben Fernandez-Casal rubenfcasal@gmail.com (ORCID)